Fix og-image endpoint for avatar
Oneblog will use either an image pulled from the issue body or the
author avatar image as an og-image for posts. When using the author
avatar, an exception would be raised because `proxyImage` was given a
string instead of a `URL`. We now coerce the `avatarUrl` into a `URL`
before passing it to `proxyImage`.